Refactor image handling out of room info into a shared component
