+++ python/nav/web/info/room/urls.py
@@ -25,12 +25,6 @@ urlpatterns = patterns('',
         name='room-info-netboxes'),
     url(r'^(?P<roomid>.+)/deviceinfo/', views.render_deviceinfo,
         name='room-info-deviceinfo'),
-    url(r'^image/update_title', views.update_title,
-        name='room-info-update-title'),
-    url(r'^image/delete', views.delete_image,
-        name='room-info-delete-image'),
-    url(r'^image/update_priority', views.update_priority,
-        name='room-info-update-priority'),
     url(r'^(?P<roomid>.+)/upload/', views.upload_image,
         name='room-info-upload'),
     url(r'^(?P<roomid>.+)/sensors/', views.render_sensors,
