+++ templates/info/room/upload.html
@@ -5,11 +5,11 @@
   <link href="{{ STATIC_URL }}css/nav/info_room.css" rel="stylesheet">
   <link href="{{ STATIC_URL }}css/nav/jquery-ui-1.8.21.custom.css"
         rel="stylesheet">
-  <script>require(["src/info_room_upload"]);</script>
+  <script>require(["src/image_upload"]);</script>
   <script>
-    NAV.urls['room-info-update-title'] = "{% url 'room-info-update-title' %}";
-    NAV.urls['room-info-delete-image'] = "{% url 'room-info-delete-image' %}";
-    NAV.urls['room-info-update-priority'] = "{% url 'room-info-update-priority' %}";
+    NAV.urls['image-update-title'] = "{% url 'image-update-title' %}";
+    NAV.urls['image-delete-image'] = "{% url 'image-delete-image' %}";
+    NAV.urls['image-update-priority'] = "{% url 'image-update-priority' %}";
   </script>
   <style>
     {# Styles for the form file input #}
@@ -49,51 +49,5 @@
 
   <p><a href="{% url 'room-info' room.id %}">Back to room</a></p>
 
-  <form id="uploadform" action="{% url 'room-info-upload' room.id %}" method="POST" enctype="multipart/form-data">
-    <input id="file" type="file" name="images" multiple class="inputfile" data-multiple-caption="{count} files selected">
-    <label for="file" class="button small"><i class="fa fa-upload"></i><span>Select images to upload</span></label>
-    <input type="submit" class="button small secondary" value="Upload selected images">
-  </form>
-
-  <div class="user-feedback">
-    {% for message in messages %}
-      <div class="alert-box {{ message.tags }}" data-alert>
-        {{ message|safe }}
-      </div>
-    {% endfor %}
-  </div>
-
-  <ul id="editimages" class="large-block-grid-4">
-  {% for image in room.image_set.all %}
-    <li class="imagecardcontainer">
-      <div data-imageid="{{ image.id }}" class="imagecard">
-        <div class="actions">
-          <i class="fa fa-times right delete" title="Delete image"></i>
-          <i class="fa fa-edit right edit" title="Edit image caption"></i>
-          <i class="fa fa-arrows right drag"
-             title="Change display order by dragging the image"></i>
-        </div>
-        <h5 class="heading">{{ image.title|safe }}</h5>
-
-        {% if image.image_exists and image.is_readable %}
-          <img src="{{ image.thumb_url }}">
-        {% else %}
-          {% if not image.image_exists %}
-            <span class="alert-box alert">Image does not exist</span>
-          {% else %}
-            <span class="alert-box alert">Image exists but is not readable</span>
-          {% endif %}
-        {% endif %}
-
-        <div class="meta">
-          <div class="left">{{ image.uploader.name }}</div>
-          <div class="right">{{ image.created }}</div>
-        </div>
-        <div class="float-clear"></div>
-      </div>
-    </li>
-
-  {% endfor %}
-  </ul>
-
+  {% include 'info/images/upload.html' %}
 {% endblock %}
